+++ webapp/components/generation/generation_server/Generation.tsx
@@ -3,7 +3,7 @@ import { Asset, AssetType, FullGeneration, FullIteration } from "@/lib/types";
 import LRPText from '@/components/generation/LRPText'
 import Image from "next/image";
 import { strToArray } from "@/lib/utils";
-import styles from '@/components/generation/Generation.module.css'
+import styles from '@/components/generation/generation_server/Generation.module.css'
 const getPriority = (item: Asset) => {
     switch (item.asset_type) {
         case AssetType.NOISE: return 1
